Server/dailydeals/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
black==26.3.1
boto3==1.35.0
botocore==1.35.99
celery==5.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.8
cryptography==46.0.7
distro==1.9.0
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.4.0
django-storages==1.14.4
django-timezone-field==7.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
easyocr==1.7.2
exceptiongroup==1.3.1
filelock==3.29.0
flower==2.0.1
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.72.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
humanize==4.15.0
idna==3.11
ImageIO==2.37.3
Jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
kombu==5.6.2
lazy-loader==0.5
MarkupSafe==3.0.3
mpmath==1.3.0
mypy_extensions==1.1.0
mysqlclient==2.2.4
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
openai==2.31.0
opencv-python==4.10.0.84
opencv-python-headless==4.13.0.92
packaging==26.0
pathspec==1.0.4
pdf2image==1.17.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
platformdirs==4.9.6
prometheus_client==0.25.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
PyMuPDF==1.27.2.2
PyMySQL==1.1.1
pyparsing==3.3.2
pypdfium2==5.7.1
pytesseract==0.3.13
python-bidi==0.6.7
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==5.0.7
requests==2.32.3
resend==2.29.0
s3transfer==0.10.4
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.4
tifffile==2025.5.10
tomli==2.4.1
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
whitenoise==6.12.0
